# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KindnessConnect API — school-safe anonymous kindness platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: schools
    description: School management
  - name: posts
    description: Posts (support and kindness acts)
  - name: comments
    description: Comments on posts
  - name: reports
    description: Content reporting
  - name: admin
    description: Admin moderation (admin role only)

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account using a school join code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with nickname and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /schools:
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a new school (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"

  /schools/{id}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get school info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: School info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          $ref: "#/components/responses/NotFound"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts for the user's school (paginated)
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [support, kindness_act]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post (text only, no images)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /posts/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post (text only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a post or comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /admin/reports:
    get:
      operationId: listAdminReports
      tags: [admin]
      summary: List reports for moderation (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, reviewed, actioned]
            default: pending
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /admin/reports/{id}:
    patch:
      operationId: updateAdminReport
      tags: [admin]
      summary: Update report status; optionally hide content or suspend user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users in the admin's school (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        joinCode:
          type: string
          description: 6-character school join code
        password:
          type: string
          minLength: 8
      required:
        - joinCode
        - password

    LoginBody:
      type: object
      properties:
        nickname:
          type: string
        schoolId:
          type: string
        password:
          type: string
      required:
        - nickname
        - schoolId
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    UserProfile:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        nickname:
          type: string
        role:
          type: string
          enum: [student, admin]
        kindnessScore:
          type: integer
        isSuspended:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - nickname
        - role
        - kindnessScore
        - isSuspended
        - createdAt

    School:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        joinCode:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - joinCode
        - isActive
        - createdAt

    CreateSchoolBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    Post:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        authorId:
          type: string
        authorNickname:
          type: string
        type:
          type: string
          enum: [support, kindness_act]
        content:
          type: string
          maxLength: 500
        isHidden:
          type: boolean
        likeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - schoolId
        - authorId
        - authorNickname
        - type
        - content
        - isHidden
        - likeCount
        - createdAt

    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit

    CreatePostBody:
      type: object
      properties:
        type:
          type: string
          enum: [support, kindness_act]
        content:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - type
        - content

    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        authorNickname:
          type: string
        content:
          type: string
          maxLength: 300
        isHidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorNickname
        - content
        - isHidden
        - createdAt

    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 300
      required:
        - content

    Report:
      type: object
      properties:
        id:
          type: string
        reporterId:
          type: string
        targetType:
          type: string
          enum: [post, comment]
        targetId:
          type: string
        reason:
          type: string
          enum: [harmful, bullying, inappropriate, spam, other]
        status:
          type: string
          enum: [pending, reviewed, actioned]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reporterId
        - targetType
        - targetId
        - reason
        - status
        - createdAt

    CreateReportBody:
      type: object
      properties:
        targetType:
          type: string
          enum: [post, comment]
        targetId:
          type: string
        reason:
          type: string
          enum: [harmful, bullying, inappropriate, spam, other]
      required:
        - targetType
        - targetId
        - reason

    UpdateReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, actioned]
        hideContent:
          type: boolean
          description: If true, hides the reported post or comment
        suspendUser:
          type: boolean
          description: If true, suspends the author of the reported content
      required:
        - status